Split Main file
