+++ fragnix.cabal
@@ -48,7 +48,7 @@ library
 
 executable fragnix
   main-is:             Main.hs
-  -- other-modules:
+  other-modules:       Build
   -- other-extensions:
   build-depends:       base >=4.6 && <4.15,
                        text >=1.2 && < 1.3,
